<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Modifications

### Before: avg 3157.98 us
<img width="1665" height="187" alt="image" src="https://github.com/user-attachments/assets/228a4132-1edf-49f1-a28e-bda7f167758e" />
<img width="2323" height="58" alt="image" src="https://github.com/user-attachments/assets/3d19560b-822e-4d7a-af22-1de656b855df" />

### After: avg 307.05 us - **10X FASTER**
<img width="1151" height="239" alt="image" src="https://github.com/user-attachments/assets/3f59312f-825a-4416-baa6-71dcfb25d65d" />
<img width="2304" height="60" alt="image" src="https://github.com/user-attachments/assets/548ddbdc-001d-4089-b176-f95264084798" />


## Accuracy Tests - GSM8k
```
python3 benchmark/gsm8k/bench_sglang.py 
```
Accuracy remains ~0.950 before and after

## Benchmark - GSM8k

| Config      | BEFORE (tokens/s) | AFTER (tokens/s) | % Faster |
|-------------|-----------------|-----------------|----------|
| TP4 MTP     | 2055.889        | 2103.981        | +2.34%   |
| TP4 DP2 MTP | 1858.580        | 1924.054        | +3.53%   |

## Checklist

- [x] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [x] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [x] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [x] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).